[remix] Create ensured dependency symlink at the `start` directory, instead of root of repo
