+++ .changeset/tricky-actors-carry.md
@@ -1,0 +1,5 @@
+---
+'@vercel/remix-builder': patch
+---
+
+Create ensured dependency symlink at the `start` directory, instead of root of repo
