+++ packages/remix/test/fixtures/00-pnpm/.gitignore
@@ -1,0 +1,1 @@
+!node_modules
