+++ packages/remix/test/fixtures/00-pnpm/package.json
@@ -1,0 +1,7 @@
+{
+  "private": true,
+  "name": "00-pnpm",
+  "dependencies": {
+    "debug": "4.3.4"
+  }
+}
